support copy file

+++ mkbootimg.sh
@@ -26,21 +26,51 @@ for file in ./boot_files/*; do
     fi
 done
 
+echo start generate gpt
+./bpttool.sh
+echo end generate gpt
 
 # start combine for u-boot
-
 UBOOT_BLOCKS=$((8 * 1024))
-BOOTLOADER_IMG=output/bootloader.img
-
-test -e ${BOOTLOADER_IMG} && rm ${BOOTLOADER_IMG}
-sync
-
-dd if=/dev/zero of=${BOOTLOADER_IMG} count=$UBOOT_BLOCKS bs=1024
-
-# gpt
-dd if=output/device-partitions.img of=${BOOTLOADER_IMG}
 # uboot
-dd if=u-boot/imx-boot-imx8qmmek-sd.bin of=${BOOTLOADER_IMG} conv=notrunc seek=32 bs=1K
+# dd if=u-boot/imx-boot-imx8qmmek-sd.bin of=${BOOTLOADER_IMG} conv=notrunc seek=32 bs=1K
 
+STD_UBOOT_FILE=imx-boot-imx8qmmek-sd.bin-flash_linux_m4
+XEN_UBOOT_FILE=u-boot-imx8qm-xen-dom0.imx
 
+for file in ./u-boot/*; do
+    if test -f ${file}; then
+        file_name=`basename ${file}`
+    
+        if [ ! "${file_name}" == "empty" ]; then
+            BOOTLOADER_IMG=
+            if [ "$file_name" == "${STD_UBOOT_FILE}" ]; then
+                BOOTLOADER_IMG=output/bootloader.img
+            elif [ "$file_name" == "${XEN_UBOOT_FILE}" ]; then
+                BOOTLOADER_IMG=output/bootloader-xen.img
+            fi
+            
+            # just support 2 type u-boot
+            if [ "${BOOTLOADER_IMG}" == "" ]; then
+                continue
+            fi
+
+            test -f ${BOOTLOADER_IMG} && rm ${BOOTLOADER_IMG}
+            sync
+            echo "${BOOTLOADER_IMG}"
+
+            dd if=/dev/zero of=${BOOTLOADER_IMG} count=$UBOOT_BLOCKS bs=1024
+            # gpt
+            dd if=output/device-partitions.img of=${BOOTLOADER_IMG}
+
+            # uboot
+            if [ "$file_name" == "${STD_UBOOT_FILE}" ]; then
+                dd if=u-boot/${STD_UBOOT_FILE} of=${BOOTLOADER_IMG} conv=notrunc seek=32 bs=1K
+            elif [ "$file_name" == "${XEN_UBOOT_FILE}" ]; then
+                dd if=u-boot/${XEN_UBOOT_FILE} of=${BOOTLOADER_IMG} conv=notrunc seek=32 bs=1K
+            fi
+        fi 
+
+    fi
+done
